requirements.txt
python-dotenv==1.0.1
moviepy==2.1.2
pillow==10.4.0
requests==2.32.3
fastapi==0.115.11
fastapi[standard]
